@@ -1,5 +1,5 @@
 import { Component, ElementRef, forwardRef, HostBinding, Inject, Injector, OnDestroy, OnInit, Optional, ViewChild, ViewEncapsulation } from '@angular/core';
-import { FormControl, FormGroup, ValidationErrors, ValidatorFn } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, ValidationErrors, ValidatorFn } from '@angular/forms';
 import { MatDialog } from '@angular/material/dialog';
 import { InputConverter } from '../../decorators/input-converter';
 import { OSafePipe } from '../../pipes/o-safe.pipe';
@@ -71,7 +71,7 @@ export class OImageComponent extends OFormDataComponent implements OnInit, OnDes
   protected _useEmptyImage: boolean = false;
   protected oSafe: OSafePipe;
   protected dialog: MatDialog;
-  public stateCtrl: FormControl;
+  public stateCtrl: UntypedFormControl;
   public src = '';
 
   constructor(
@@ -132,7 +132,7 @@ export class OImageComponent extends OFormDataComponent implements OnInit, OnDes
 
   public createFormControl(cfg?: { value: any, disabled: boolean }, validators?: ValidatorFn[]): OFormControl {
     this._fControl = super.createFormControl(cfg, validators);
-    this.stateCtrl = new FormControl(void 0, this.resolveValidators());
+    this.stateCtrl = new UntypedFormControl(void 0, this.resolveValidators());
     this._fControl.fControlChildren = [this.stateCtrl];
     return this._fControl;
   }
@@ -223,10 +223,10 @@ export class OImageComponent extends OFormDataComponent implements OnInit, OnDes
     return this._useEmptyImage && this.isEmpty();
   }
 
-  public getFormGroup(): FormGroup {
-    let formGroup: FormGroup = super.getFormGroup();
+  public getFormGroup(): UntypedFormGroup {
+    let formGroup: UntypedFormGroup = super.getFormGroup();
     if (!formGroup) {
-      formGroup = new FormGroup({});
+      formGroup = new UntypedFormGroup({});
       formGroup.addControl(this.getAttribute(), this.getControl());
     }
     return formGroup;
@@ -273,7 +273,7 @@ export class OImageComponent extends OFormDataComponent implements OnInit, OnDes
     }
   }
 
-  protected maxFileSizeValidator(control: FormControl): ValidationErrors {
+  protected maxFileSizeValidator(control: UntypedFormControl): ValidationErrors {
     if (control.value && control.value.length > 0 && Util.isDefined(this.maxFileSize)) {
       if (!Util.isDefined(this.fileInput.nativeElement.files)) {
         return {};
