@@ -10,7 +10,7 @@ import {
   ViewChild,
   ViewEncapsulation,
 } from '@angular/core';
-import { FormControl, NG_VALUE_ACCESSOR } from '@angular/forms';
+import { UntypedFormControl, NG_VALUE_ACCESSOR } from '@angular/forms';
 import moment from 'moment'
 
 import { InputConverter } from '../../../decorators/input-converter';
@@ -40,7 +40,7 @@ export class DaterangepickerComponent implements OnInit {
   chosenLabel: string;
   calendarVariables: { left: any, right: any } = { left: {}, right: {} };
   timepickerVariables: { left: any, right: any } = { left: {}, right: {} };
-  daterangepicker: { start: FormControl, end: FormControl } = { start: new FormControl(), end: new FormControl() };
+  daterangepicker: { start: UntypedFormControl, end: UntypedFormControl } = { start: new UntypedFormControl(), end: new UntypedFormControl() };
   applyBtn: { disabled: boolean } = { disabled: false };
   startDate = moment().startOf('day');
   endDate = moment().endOf('day');
