@@ -1,5 +1,5 @@
 import { ChangeDetectionStrategy, Component, Inject, Injector } from '@angular/core';
-import { FormControl, FormGroup, Validators } from '@angular/forms';
+import { UntypedFormControl, UntypedFormGroup, Validators } from '@angular/forms';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
 
 import { OTableBaseDialogClass } from '../../../../components/table/extensions/dialog/o-table-base-dialog.class';
@@ -13,12 +13,12 @@ import { OFilterDefinition } from '../../../../types/o-filter-definition.type';
 export class OStoreFilterDialogComponent extends OTableBaseDialogClass {
 
   filterNames: Array<string> = [];
-  formGroup: FormGroup = new FormGroup({
-    name: new FormControl('', [
+  formGroup: UntypedFormGroup = new UntypedFormGroup({
+    name: new UntypedFormControl('', [
       Validators.required,
       this.filterNameValidator.bind(this)
     ]),
-    description: new FormControl('')
+    description: new UntypedFormControl('')
   });
 
   constructor(
@@ -39,7 +39,7 @@ export class OStoreFilterDialogComponent extends OTableBaseDialogClass {
     return this.formGroup.value;
   }
 
-  protected filterNameValidator(control: FormControl) {
+  protected filterNameValidator(control: UntypedFormControl) {
     const ctrlValue: string = control.value;
     if (this.filterNames.indexOf(ctrlValue) !== -1) {
       return { filterNameAlreadyExists: true };
